refactor : todoApp 파일경로 정리
