+++ index.html
@@ -8,7 +8,7 @@
       href="https://fonts.googleapis.com/icon?family=Material+Icons"
       rel="stylesheet"
     />
-    <script type="module" src="src/main.js"></script>
+    <script type="module" src="src/index.js"></script>
   </head>
   <body>
     <div id="app">
