+++ src/index.js
@@ -1,7 +1,6 @@
-const onUserCreateHandler = () => {
-  const userName = prompt("추가하고 싶은 이름을 입력해주세요.");
-}
+'use strict';
 
-const userCreateButton = document.querySelector('.user-create-button')
-userCreateButton.addEventListener('click', onUserCreateHandler)
+import TodoApp from './todoApp.js';
 
+const todoApp = new TodoApp();
+todoApp.init();
